Started layout revision with Tailwind css
