+++ IDEAS.txt
@@ -1,0 +1,21 @@
+Features - 
+1. Searching
+2. Gamification 
+    - Add fun elements such as a points system or leaderboards. Inorder to get customers to make more purchases.
+    - Maybe check if the user hasnt been on the website in a little while it will show them a message to get on the leaderboards and get more points.
+3. Automatic Reordering (DONE)
+    - Make a user page so that users can see previous orders and reorder them.
+4. User page (DONE)
+    - add a basic user page with previous orders.
+5. Loyalty program
+    - The points that people earn allow them to spend them for discounts, or free products, or other perks (free shipping)
+6. Birthday rewards
+    - would need to redesign the registration page
+
+
+Task - 
+1. Admin page
+    - Will allow the admin(s) to create discounts, edit the leaderboards. View all registered users and edit their profiles.
+
+QOL feature -
+1. Make it so when you add an item to cart it stays on the product page.
