+++ app/static/css/input.css
@@ -1,0 +1,50 @@
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
+
+@keyframes slideIn {
+    0% {
+        transform: scale(0.25);
+        opacity: 0;
+        -webkit-transform: scale(0.25);
+        -moz-transform: scale(0.25);
+        -ms-transform: scale(0.25);
+        -o-transform: scale(0.25);
+    }
+
+    100% {
+        transform: scale(1);
+        opacity: 1;
+    }
+}
+
+@keyframes slideOut {
+    0% {
+        transform: scale(1);
+        opacity: 1;
+    }
+
+    99% {
+        opacity: 0;
+    }
+
+    100% {
+        transform: scale(0.25);
+        opacity: 0;
+        -webkit-transform: scale(0.25);
+        -moz-transform: scale(0.25);
+        -ms-transform: scale(0.25);
+        -o-transform: scale(0.25);
+    }
+}
+
+* {
+    font-family: "Rubik", sans-serif;
+}
+
+.bg-image {
+    background-image: url(../img/uuunion.svg);
+    background-repeat: no-repeat;
+    background-size: cover;
+    background-position: center;
+}
