+++ app/templates/account.html
@@ -1,0 +1,77 @@
+{% extends "base.html" %} {% block content %}
+<main class="bg-gray-100 ">
+    <div class="container m-auto flex py-4">
+        <ul class="flex grow flex-col gap-5">
+            <li>
+                <i class="fa fa-user text-2xl mr-3"></i>
+                <a href="" class="text-2xl">Account</a>
+            </li>
+            <li>
+                <i class="fa fa-bookmark text-2xl mr-3"></i>
+                <a href="" class="text-2xl">My Orders</a>
+            </li>
+            <li>
+                <i class="fa fa-lock text-2xl mr-3"></i>
+                <a href="" class="text-2xl">Password</a>
+            </li>
+        </ul>
+        <div class="flex grow-[2] flex-col bg-white p-5 rounded-md shadow-md">
+            <h1>Account Details</h1>
+            <p>Update your account detail.</p>
+            <form action="" class="">
+                <div class="flex items-center gap-5">
+                    <div class="w-1/3 flex flex-col">
+                        <label for="name">First Name:</label>
+                        <input type="email" name="email" id="email" placeholder="johndoe@gmail.com"
+                            class="w-full mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200">
+                    </div>
+                    <div class="w-1/3  flex flex-col">
+                        <label for="name">Last Name:</label>
+                        <input type="text" name="name" id="name" placeholder="John Doe"
+                            class="w-full mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200">
+                    </div>
+                    <div class="w-1/3 flex flex-col">
+                        <label for="name">Email:</label>
+                        <input type="email" name="email" id="email" placeholder="johndoe@gmail.com"
+                            class="w-full mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200">
+                    </div>
+                </div>
+                <div class="flex items-center gap-5">
+                    <div class="w-1/2  flex flex-col">
+                        <label for="name">Address:</label>
+                        <input type="text" name="name" id="name" placeholder="John Doe"
+                            class="w-full mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200">
+                    </div>
+                    <div class="w-1/4  flex flex-col">
+                        <label for="name">City:</label>
+                        <input type="text" name="name" id="name" placeholder="John Doe"
+                            class="w-full mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200">
+                    </div>
+                    <div class="w-1/4  flex flex-col">
+                        <label for="name">State:</label>
+                        <input type="text" name="name" id="name" placeholder="John Doe"
+                            class="w-full mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200">
+                    </div>
+                </div>
+                <div class="flex items-center gap-5">
+                    <div class="w-1/2  flex flex-col">
+                        <label for="name">Country:</label>
+                        <input type="text" name="name" id="name" placeholder="John Doe"
+                            class="w-full mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200">
+                    </div>
+                    <div class="w-1/2  flex flex-col">
+                        <label for="name">Zip Code:</label>
+                        <input type="text" name="name" id="name" placeholder="John Doe"
+                            class="w-full mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200">
+                    </div>
+                    <div class="w-1/2  flex flex-col">
+                        <label for="name">Phone:</label>
+                        <input type="text" name="name" id="name" placeholder="John Doe"
+                            class="w-full mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200">
+                    </div>
+                </div>
+            </form>
+        </div>
+    </div>
+</main>
+{% endblock %}
