+++ app/templates/checkout.html
@@ -1,0 +1,148 @@
+{% extends "base.html" %} {% block content %}
+<div class="checkout-container">
+    <div class="container my-5 py-5 ">
+        <div class="row">
+            <div class="col-md-8 w-50">
+                <h2>Shipping Information</h2>
+                <form action="/process_checkout" method="post">
+                    <div class="row">
+                        <div class="col-md-6 mb-3">
+                            <label for="first_name">First Name</label>
+                            <input
+                                type="text"
+                                class="form-control"
+                                id="first_name"
+                                name="first_name"
+                                value="{{user.first_name or ''}}"
+                                required />
+                        </div>
+                        <div class="col-md-6 mb-3">
+                            <label for="last_name">Last Name</label>
+                            <input
+                                type="text"
+                                class="form-control"
+                                id="last_name"
+                                name="last_name"
+                                value="{{user.last_name or ''}}"
+                                required />
+                        </div>
+                    </div>
+                    <div class="mb-3">
+                        <label for="email">Email Address</label>
+                        <input
+                            type="email"
+                            class="form-control"
+                            id="email"
+                            name="email"
+                            value="{{user.email or ''}}"
+                            required />
+                    </div>
+                    <div class="mb-3">
+                        <label for="address">Address</label>
+                        <input
+                            type="text"
+                            class="form-control"
+                            id="address"
+                            name="address"
+                            value="{{user.address or ''}}"
+                            required />
+                    </div>
+                    <div class="row">
+                        <div class="col-md-4 mb-3">
+                            <label for="city">City</label>
+                            <input
+                                type="text"
+                                class="form-control"
+                                id="city"
+                                name="city"
+                                value="{{user.city or ''}}"
+                                required />
+                        </div>
+                        <div class="col-md-4 mb-3">
+                            <label for="state">State</label>
+                            <input
+                                type="text"
+                                class="form-control"
+                                id="state"
+                                name="state"
+                                value="{{user.state or ''}}"
+                                required />
+                        </div>
+                        <div class="col-md-4 mb-3">
+                            <label for="zip_code">Zip Code</label>
+                            <input
+                                type="text"
+                                class="form-control"
+                                id="zip_code"
+                                name="zip_code"
+                                value="{{user.zip_code or ''}}"
+                                required />
+                        </div>
+                    </div>
+                    <div class="mb-3">
+                        <label for="country">Country</label>
+                        <input
+                            type="text"
+                            class="form-control"
+                            id="country"
+                            name="country"
+                            value="{{user.country or ''}}"
+                            required />
+                    </div>
+                    <div class="mb-3">
+                        <label for="phone">Phone Number</label>
+                        <input
+                            type="tel"
+                            class="form-control"
+                            id="phone"
+                            name="phone"
+                            value="{{user.phone or ''}}"
+                            required />
+                    </div>
+                    <button
+                        type="submit"
+                        class="btn btn-primary fit-content mt-3 p-2">
+                        Checkout
+                    </button>
+                </form>
+                <hr />
+                <h2>Payment Details (Not integrated)</h2>
+                <form>
+                    <!-- Payment information form fields go here -->
+                </form>
+            </div>
+            <div class="col-md-4 w-50">
+                <div class="card">
+                    <div class="card-header">
+                        <h3>Cart Summary</h3>
+                    </div>
+                    <div class="card-body">
+                        <!-- Loop through cart items and display them -->
+                        <ul class="list-group list-group-flush">
+                            {% for item in order_data %}
+                            <li
+                                class="list-group-item d-flex justify-content-between align-items-center gap-4">
+                                <img src="{{item.image}}" alt="" class="w-15" />
+                                {{ item.title }} ({{ item.quantity }})
+
+                                <span>${{item.price}}</span>
+                                <a href="/remove_item/{{item.product_id}}"
+                                    >Remove</a
+                                >
+                            </li>
+                            {% endfor %}
+                        </ul>
+                    </div>
+                    <div class="card-footer">
+                        <h4
+                            class="d-flex justify-content-between align-items-center">
+                            Total
+                            <span>${{ cart_total or '0.00'}}</span>
+                        </h4>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+{% endblock %}
