+++ app/templates/order_details.html
@@ -1,0 +1,43 @@
+{% extends "base.html" %}
+{% block content %}
+<div class="container mb-2">
+    <div class="row">
+        <div class="col-md-12">
+            <div class="card">
+                <div class="card-body">
+                    <h5 class="card-title">Order Details</h5>
+                    <p class="card-text"><strong>Order ID:</strong> {{order_id}}</p>
+                    <p class="card-text"><strong>Order Date:</strong> {{ order.created_at}}</p>
+
+                    <table class="table table-hover my-3">
+                        <thead>
+                            <tr>
+                                <th scope="col">Image</th>
+                                <th scope="col">Title</th>
+                                <th scope="col">Description</th>
+                                <th scope="col">Quantity</th>
+                                <th scope="col">Price</th>
+                            </tr>
+                        </thead>
+                        <tbody>
+                            {% for item in orders %}
+                            <tr>
+                                <td>
+                                    <img src="{{ item.image }}" alt="{{ item.title }}" width="50">
+                                </td>
+                                <td class="fw-bolder">{{ item.title }}</td>
+                                <td>{{ item.description }}</td>
+                                <td>{{ item.quantity }}</td>
+                                <td>${{ item.price }}</td>
+                            </tr>
+                            {% endfor %}
+                            
+                        </tbody>
+                    </table>
+                    <p class="card-text"><strong>Total Price:</strong> ${{ order.total_price }}</p>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+{% endblock %}
