+++ app/templates/register.html
@@ -1,0 +1,74 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.7.0/css/font-awesome.min.css" />
+    <link rel="preconnect" href="https://fonts.googleapis.com" />
+    <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
+    <link href="https://fonts.googleapis.com/css2?family=Rubik&display=swap" rel="stylesheet">
+
+    <link href="{{url_for('static', filename='css/output.css')}}" rel="stylesheet">
+    <script defer type="text/javascript" src="{{url_for('static', filename='js/script.js')}}"></script>
+    <title>Home</title>
+</head>
+
+<body class="bg-image">
+    <main class="flex items-center justify-center h-screen flex-col ">
+        <a href="/" class="flex items-center mb-10">
+            <img src="{{url_for('static', filename='img/icon.png')}}" alt="" class="w-14 mr-3">
+            <h1 class="text-3xl font-bold text-color ">Groovy Gear</h1>
+        </a>
+
+        <form action="/register" method="post" class="w-[400px]">
+            {% with messages = get_flashed_messages() %}
+            {% if messages %}
+            {% for message in messages %}
+            <p>{{message}}</p>
+            {% endfor %}
+            {% endif %}
+            {% endwith %}
+            <div class="flex flex-col my-5">
+                <label for="first_name" class="font-bold">First Name:</label>
+                <input type="text" name="first_name" id="first_name" placeholder="John"
+                    class="mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200"
+                    required>
+            </div>
+            <div class="flex flex-col mb-5">
+                <label for="last_name" class="font-bold">Last Name:</label>
+                <input type="text" name="last_name" id="last_name" placeholder="Doe"
+                    class="mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200"
+                    required>
+            </div>
+
+            <div class="flex flex-col mb-5">
+                <label for="email" class="font-bold">Email address:</label>
+                <input type="email" name="email" id="email" placeholder="johndoe@gmail.com"
+                    class="mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200"
+                    required>
+            </div>
+            <div class="flex flex-col mb-5">
+                <label for="password" class="font-bold">Password:</label>
+                <input type="password" name="password" id="password" placeholder="Abcd123!@#"
+                    class="mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200"
+                    required>
+            </div>
+            <div class="flex flex-col mb-5">
+                <label for="confirm_password" class="font-bold">Confirm Password:</label>
+                <input type="password" name="confirm_password" id="confirm_password" placeholder="Abcd123!@#"
+                    class="mt-2 text-slate-900 bg-white rounded-md px-3 h-10 shadow-md focus:outline-none focus:ring-2 focus:ring-groovy-purple ring-1 ring-slate-200"
+                    required>
+            </div>
+            <button type="submit" class="py-2 bg-groovy-red rounded-md w-full shadow-md text-bold text-white">Sign up
+                for an account</button>
+        </form>
+        <p class="pt-5">Already have an account? <a href="/login" class="text-white underline hover:text-gray-200">Login!</a>
+        </p>
+
+    </main>
+
+</body>
+
+</html>
