+++ app/templates/show_product.html
@@ -1,0 +1,26 @@
+{% extends "base.html" %} {% block content %}
+
+<div class="container d-flex justify-content-between align-items-center height">
+    <div class="w-40">
+        <img
+            src="{{product.image}}"
+            alt=""
+            class="img-fluid ${3|rounded-top,rounded-right,rounded-bottom,rounded-left,rounded-circle,|} hover" />
+    </div>
+    <div class="w-40">
+        <h1>{{product.title}}</h1>
+        <p>{{product.description}}</p>
+        <p class="fs-2 fw-bolder">${{product.price}}</p>
+        <form action="/add_to_cart" method="post">
+            <input type="hidden" name="product_id" value="{{product.id}}" />
+            <input
+                type="hidden"
+                name="product_price"
+                value="{{product.price}}" />
+            <input type="hidden" name="product_quantity" value="1" />
+            <button type="submit" class="btn btn-primary button-hover">Add to Cart</button>
+        </form>
+    </div>
+</div>
+
+{% endblock%}
