--- flask_app/static/css/style.css
@@ -1,149 +1,0 @@
-@keyframes slideIn {
-    0% {
-        transform: scale(0.25);
-        opacity: 0;
-        -webkit-transform: scale(0.25);
-        -moz-transform: scale(0.25);
-        -ms-transform: scale(0.25);
-        -o-transform: scale(0.25);
-    }
-
-    100% {
-        transform: scale(1);
-        opacity: 1;
-    }
-}
-
-@keyframes slideOut {
-    0% {
-        transform: scale(1);
-        opacity: 1;
-    }
-
-    99% {
-        opacity: 0;
-    }
-
-    100% {
-        transform: scale(0.25);
-        opacity: 0;
-        -webkit-transform: scale(0.25);
-        -moz-transform: scale(0.25);
-        -ms-transform: scale(0.25);
-        -o-transform: scale(0.25);
-    }
-}
-
-* {
-    font-family: "Nunito", sans-serif;
-    margin: 0;
-    padding: 0;
-}
-.checkout-container, .container {
-    padding-top: 5rem; /* Add extra padding to the top to avoid overlapping with the navbar */
-}
-.exclusive {
-    height: 350px;
-}
-
-.product {
-    width: 20%;
-}
-.hide-element {
-    display: none;
-}
-.slide-element {
-    animation: slideIn 0.35s ease-in-out;
-    -webkit-animation: slideIn 0.35s ease-in-out;
-}
-.slide-out-element {
-    animation: slideOut 0.35s ease-in-out;
-    -webkit-animation: slideOut 0.35s ease-in-out;
-}
-.slideshow-container,
-.height {
-    height: 100vh;
-    
-}
-.w-33 {
-    width: 33%;
-}
-.w-40 {
-    width: 40%;
-}
-.w-15 {
-    width: 15%;
-}
-.fit-content {
-    width: fit-content;
-    height: fit-content;
-}
-
-h2 {
-    width: fit-content;
-}
-
-.arrow-left {
-    margin-left: 75px;
-}
-.arrow-right {
-    margin-right: 75px;
-}
-h2,
-.navbar-brand {
-    font-weight: 800;
-}
-.clickable {
-    cursor: pointer;
-}
-
-.hover {
-    transition: 0.5s;
-    transform: scale(1);
-    -webkit-transform: scale(1);
-    -moz-transform: scale(1);
-    -ms-transform: scale(1);
-    -o-transform: scale(1);
-    -webkit-transition: 0.5s;
-    -moz-transition: 0.5s;
-    -ms-transition: 0.5s;
-    -o-transition: 0.5s;
-}
-.button-hover {
-    transition: 0.5s;
-    transform: scale(1);
-}
-.button-hover:hover {
-    transition: 0.5s;
-    transform: scale(1.12);
-}
-.hover:hover {
-    transition: 0.5s;
-    transform: scale(1.02) rotate(2deg);
-    -webkit-transform: scale(1.02) rotate(2deg);
-    -moz-transform: scale(1.02) rotate(2deg);
-    -ms-transform: scale(1.02) rotate(2deg);
-    -o-transform: scale(1.02) rotate(2deg);
-    -webkit-transition: 0.5s;
-    -moz-transition: 0.5s;
-    -ms-transition: 0.5s;
-    -o-transition: 0.5s;
-}
-
-.bg-darker-custom {
-    background-color: #f5f6f6;
-}
-body {
-    background-color: #ffffff;
-}
-.button-yellow {
-    background-color: #CCCCCC;
-}
-.alert-container {
-    z-index: 2000;
-    background-color: rgba(0, 0, 0, 0.5);
-    backdrop-filter: blur(3px);
-}
-.alert-card {
-    max-width: 600px;
-}
