--- flask_app/templates/login.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-{% extends "base.html" %}
-
-{% block content %}
-
-<div class="d-flex justify-content-center align-items-center height">
-    <form action="/login" method="post" class=" rounded-2 shadow-lg w-25 p-4 d-flex flex-column gap-3 ">
-
-        <div class="form-group">
-            <label for="email">Email:</label>
-            <input type="email" class="form-control" name="email" id="email" aria-describedby="emailHelpId"
-                placeholder="johndoe@gmail.com">
-        </div>
-        <div class="form-group">
-            <label for="password">Password:</label>
-            <input type="password" class="form-control" name="password" id="password" placeholder="Abcd123!@#">
-        </div>
-        <button type="submit" class="btn btn-primary  w-25 m-auto">Login</button>
-        <h4 class="text-center">Don't have an account? <a name="" id="" class="fs-4 " href="/register"
-                role="button">Click
-                here to Sign Up!</a></h4>
-    </form>
-</div>
-
-{% endblock %}
