+++ server.py
@@ -1,8 +1,8 @@
-from flask_app.controllers import *
-from flask_app import app
+from app.controllers import *
+from app import app
 from flask import Flask, render_template, session
-from flask_app.models import product, user
-from flask_app.controllers import order_item, users, orders, products
+from app.models import product, user
+from app.controllers import order_item, users, orders, products
